Add support for navigating a todo lists history

+++ orleans/JournaledTodoList/JournaledTodoList.WebApp/Components/Pages/TodoListPage.razor.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Immutable;
+using System.Diagnostics.CodeAnalysis;
 using JournaledTodoList.WebApp.Grains;
 using JournaledTodoList.WebApp.Grains.Events;
 using JournaledTodoList.WebApp.Services;
@@ -11,6 +12,10 @@ public partial class TodoListPage(TodoListService todoService)
     private string newItemTitle = "";
     private TodoList? todoList;
     private ImmutableArray<TodoListEvent> history;
+    private DateTimeOffset? currentViewTimestamp;
+
+    [MemberNotNullWhen(true, nameof(currentViewTimestamp))]
+    private bool IsViewingHistory => !history.IsDefaultOrEmpty && currentViewTimestamp < history[^1].Timestamp;
 
     [Parameter, EditorRequired]
     public required string ListId { get; set; }
@@ -19,6 +24,7 @@ public partial class TodoListPage(TodoListService todoService)
     {
         if (todoList?.Name != ListId)
         {
+            currentViewTimestamp = null;
             todoList = null;
             await LoadTodoList();
         }
@@ -32,7 +38,7 @@ public partial class TodoListPage(TodoListService todoService)
 
     private async Task AddItem()
     {
-        if (string.IsNullOrWhiteSpace(newItemTitle))
+        if (string.IsNullOrWhiteSpace(newItemTitle) || IsViewingHistory)
         {
             return;
         }
@@ -44,7 +50,7 @@ public partial class TodoListPage(TodoListService todoService)
 
     private async Task UpdateItem(int itemId, string? title)
     {
-        if (string.IsNullOrWhiteSpace(title))
+        if (string.IsNullOrWhiteSpace(title) || IsViewingHistory)
         {
             return;
         }
@@ -55,13 +61,47 @@ public partial class TodoListPage(TodoListService todoService)
 
     private async Task ToggleItem(int itemId)
     {
+        if (IsViewingHistory)
+        {
+            return;
+        }
+
         await todoService.ToggleTodoItemAsync(ListId, itemId);
         await LoadTodoList();
     }
 
     private async Task RemoveItem(int itemId)
     {
+        if (IsViewingHistory)
+        {
+            return;
+        }
+
         await todoService.RemoveTodoItemAsync(ListId, itemId);
         await LoadTodoList();
     }
+
+    private async Task ViewAtTimestamp(DateTimeOffset timestamp)
+    {
+        if (timestamp == history[^1].Timestamp)
+        {
+            currentViewTimestamp = null;
+        }
+        else
+        {
+            currentViewTimestamp = timestamp;
+            todoList = await todoService.GetTodoListAtTimestampAsync(ListId, timestamp);
+        }
+    }
+
+    private async Task ReturnToCurrentVersion()
+    {
+        currentViewTimestamp = null;
+        await LoadTodoList();
+    }
+
+    private bool IsCurrentHistoryItem(TodoListEvent item)
+        => currentViewTimestamp.HasValue
+        ? item.Timestamp == currentViewTimestamp
+        : history[^1] == item;
 }
